pyproject.toml
[tool.poetry]
name = "trusted-routes"
version = "0.1.0"
description = "Backend CRM API for travel app"
authors = ["krumitz <aaronbahmid@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
gunicorn = "^23.0.0"
djangorestframework = "^3.15.2"
django-storages = "^1.14.4"
requests = "^2.32.3"
python-decouple = "^3.8"
redis = "^5.2.0"
drf-access-policy = "^1.5.0"
drf-extra-fields = "^3.7.0"
drf-standardized-errors = "^0.14.1"
drf-yasg = "^1.21.8"
s3transfer = "^0.10.4"
boto3 = "^1.35.74"
whitenoise = "^6.8.2"
django-rest-knox = "^5.0.2"
django-oauth-toolkit = "^3.0.1"
django-environ = "^0.11.2"
django-cors-headers = "^3.13.0"
django-cors-middleware = "^1.5.0"
coreapi = "^2.3.3"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
